+++ src/lib/model/ticket.ts
@@ -15,6 +15,9 @@
  * To create a Ticket, call {@link TicketService.create}.
  * To search for tickets, call {@link TicketService.search}.
  */
+import { TicketExtra } from './ticket-extra';
+import { TicketLabel } from './ticket-label';
+
 export interface Ticket {
   /**
    * This ticket's unique ID. For example, 14995020
@@ -154,52 +157,6 @@ export interface Ticket {
 }
 
 /**
- * Represents a custom field attached to a Ticket.
- * Users can define custom fields to show in the Service Screen. Those custom fields will be
- * stored in ticket extras.
- *
- * The title of a TicketExtra is the heading shown above the input field.
- * The value of a TicketExtra is the string typed into the input field, or the string value
- * selected in the dropdown. For multiselect fields, multiple TicketExtras are generated, one
- * for each value.
- *
- * For example, if the Location has an input field "Order ID" with the Text type, the
- * TicketExtra would look like this:
- *
- * ```
- * { "title": "Order ID", "value": "AB1234CDEF" }
- * ```
- *
- * If the Location has a Select input field named "Favorite Greek Letter" that has two options,
- * "Alpha" and "Beta", and "Beta" is selected, then the TicketExtra looks like this:
- *
- * ```
- * { "title": "Favorite Greek Letter", "value": "Alpha" }
- * ```
- *
- * If both Alpha and Beta are selected, then two TicketExtras are created:
- *
- * ```
- * { "title": "Favorite Greek Letter", "value": "Alpha" },
- * { "title": "Favorite Greek Letter", "value": "Beta" }
- * ```
- *
- * If the custom field is an hyperlink, then the URL of the link is stored in the "url"
- * property, and the title is always assigned to "Open".
- *
- * For example, for a custom hyperlink called "Twitter Page", this would be:
- *
- * ```
- * { "title": "Twitter Page", "value": "Open", "url": "https://www.twitter.com/" }
- * ```
- */
-export interface TicketExtra {
-  title: string;
-  value: string;
-  url?: string;
-}
-
-/**
  * Represents an Interaction attached to a ticket.
  *
  * Interactions log the various steps the visitor takes to proceed from start of service to the end.
@@ -215,21 +172,6 @@ export interface TicketInteraction {
   start: string;
   end: string;
   line: number;
-}
-
-/**
- * Represents a ticket label.
- *
- * Tickets can have one or more labels, each identified by their 'value' string. The color of a
- * label is automatically generated.
- *
- * The color of a label is a 24-bit RGB hex color, without the hash mark (#) prefix.
- *
- * For example: { "value": "Has documents", "color": "00FF00" }
- */
-export interface TicketLabel {
-  value: string;
-  color: string;
 }
 
 /**
